+++ R/estimate.BSVARMSH.R
@@ -7,7 +7,7 @@
 #' for the structural matrix \eqn{B} and the equation-by-equation sampler by Chan, Koop, & Yu (2021)
 #' for the autoregressive slope parameters \eqn{A}. Additionally, the parameter matrices \eqn{A} and \eqn{B}
 #' follow a Minnesota prior and generalised-normal prior distributions respectively with the matrix-specific
-#' overall shrinkage parameters estimated thanks to a 3-level hierarchical prior distribution. The MS
+#' overall shrinkage parameters estimated thanks to a hierarchical prior distribution. The MS
 #' model is estimated using the prior distributions and algorithms proposed by Woźniak & Droumaguet (2022).
 #' See section \bold{Details} for the model equations.
 #'